Add AI-powered arrival time recommendations and comprehensive social features
Major Features Added:

1. AI Arrival Time Recommendations
   - New API endpoint using GPT-4o to analyze conditions
   - Analyzes 8+ data sources (snow, weather, roads, crowds, lifts)
   - Provides optimal arrival time with confidence levels
   - Shows alternatives and pro tips
   - Integrated into Overview and Travel tabs

2. Tabbed Mountain View Interface
   - Reorganized mountain details into 8 categorized tabs
   - Overview, Forecast, History, Travel, Safety, Webcams, Social, Lifts
   - Smooth tab switching with color-coded icons
   - Fully implemented Overview and Forecast tabs
   - Stub implementations for remaining tabs

3. Social Features Backend
   - Supabase authentication integration
   - Photo upload with S3 storage
   - Comments system with nested replies
   - Check-ins and trip reports
   - Likes for photos, comments, and check-ins
   - Push notification infrastructure

4. iOS Social Features
   - Sign in with Apple integration
   - Photo upload and gallery views
   - Comment list and input views
   - Check-in forms and displays
   - Like button interactions
   - Alert subscriptions management

5. Push Notifications
   - Apple Push Notification service setup
   - Device token registration
   - Weather and powder alert system
   - Cron jobs for alert checking

6. Bug Fixes & Improvements
   - Fixed TabButton and StatCard name collisions
   - Fixed ForecastTab type mismatches
   - Removed duplicate model definitions
   - Updated Swift version from 6.0 to 5.0
   - Fixed all PushNotificationManager warnings
   - Improved error handling across all services

Technical Details:
- Backend: Next.js API routes with OpenAI GPT-4o integration
- iOS: SwiftUI with Observation framework
- Database: Supabase (PostgreSQL + Storage)
- Testing: Added Vitest configuration and unit tests

Files: 100+ new files, 2400+ insertions

+++ ios/PowderTracker/PowderTracker/Models/SnowComparison.swift
@@ -22,7 +22,7 @@ struct YearOverYearComparison: Codable {
 
 struct SnowDepthDataPoint: Codable {
     let date: String
-    let snowDepth: Double
+    let snowDepth: Int
 }
 
 struct BaseDepthGuidelines: Codable {
